+++ requirements.txt
@@ -1,3 +1,1 @@
 flask
-PyMongo
-Pillow
