+++ web/config.py
@@ -9,7 +9,3 @@ class Config:
     SECRET_KEY = os.environ['SECRET_KEY']
     DEBUG = os.environ['FLASK_ENV'] == 'development'
 
-    MONGO_HOST = os.environ['MONGO_HOST']
-    MONGO_PORT = os.environ['MONGO_PORT']
-    MONGO_INITDB_ROOT_USERNAME = os.environ['MONGO_INITDB_ROOT_USERNAME']
-    MONGO_INITDB_ROOT_PASSWORD = os.environ['MONGO_INITDB_ROOT_PASSWORD']
