--- web/static/style.css
@@ -1,164 +1,0 @@
-
-/*=============================== Base ===============================*/
-body {
-    font-family: 'ProximaNova-Regular','HelveticaNeue','Helvetica Neue','HelveticaNeueRoman','HelveticaNeue-Roman','Helvetica Neue Roman','TeXGyreHerosRegular','Helvetica','Tahoma','Geneva','Arial',sans-serif;
-    width: 70%;
-    max-width: 1000px;
-    min-width: 500px;
-    margin: auto;
-}
-
-
-
-ul {
-    list-style-type: none;
-}
-
-header {
-
-}
-
-.header-h1 {
-    margin: 0;
-}
-
-
-nav {
-
-}
-
-.topBar {
-    display: flex;
-    height: 50px;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.nav-ul {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.nav-li {
-    display: block;
-}
-
-.nav-li-a {
-    display: block;
-    width: 60px;
-    text-align: center;
-}
-
-
-
-/*=============================== Login Overlay ===============================*/
-
-
-#login-overlay {
-    width: 300px;
-    height: 100px;
-    position: fixed;
-    top: 30%;
-    padding: 50px;
-    left: calc(50% - 150px);
-    background: white;
-    border-width: 1px;
-    border-style: solid;
-    text-align: center;
-}
-
-#login-overlay-close {
-    position: absolute;
-    right: 10px;
-    top: 10px;
-}
-
-#login-overlay-form {
-
-}
-
-
-/*=============================== Search ===============================*/
-
-.search {
-    height: 40px;
-}
-
-.search-bar {
-    height: 30px;
-    max-width: 600px;
-    min-width: 500px;
-    margin: 10px auto 0;
-    display: flex;
-}
-
-#search-type {
-    -webkit-appearance: none;
-    width: 100px;
-    height: 30px;
-    border-radius: 5px 0 0 5px;
-    border: 2px solid lightgray;
-    border-right-width: 0;
-    padding-left: 5px;
-    background-position: right 50%;
-    background-repeat: no-repeat;
-    background-image: url('data:image/svg+xml;utf8,<?xml version="1.0" encoding="utf-8"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd"><svg xmlns="http://www.w3.org/2000/svg" width="14" height="12" version="1"><path d="M4 8L0 4h8z"/></svg>');
-}
-
-#search-form {
-    width: calc(100% - 100px);
-    display: flex;
-}
-
-#search-box {
-    width: calc(100% - 60px);
-    height: 30px;
-    border: 2px solid lightgray;
-    border-right-width: 0;
-}
-
-#search-submit {
-    width: 80px;
-    height: 30px;
-    border: 2px solid lightgray;
-    background-color: whitesmoke;
-    border-radius: 0 5px 5px 0;
-}
-
-#suggestions {
-    position: relative;
-    width: 500px;
-    margin: auto;
-    background: white;
-}
-
-.result-overview {
-    color: grey;
-}
-
-
-
-
-
-
-/*=============================== Documents ===============================*/
-
-.document {
-
-}
-
-.document-block {
-    margin-bottom: 20px;
-}
-
-.document-overview {
-    width: 100%;
-}
-
-.document-keyFeatures {
-    list-style-type: circle;
-}
-
-.document-keyApplications {
-    list-style-type: circle;
-}
