--- web/templates/demo.html
@@ -1,133 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <title>Demo - IME Inventory DB</title>
-    <link rel="shortcut icon" href="{{ url_for('static', filename='assets/favicon.ico') }}">
-    <link rel="stylesheet" href="{{ url_for('static', filename='demo.css') }}">
-    {% block script %}{% endblock %}
-</head>
-<body>
-
-<section class="navbar">
-    <div class="navbar-link__wrapper"><a href="#">About</a></div>
-    <div class="navbar-link__wrapper"><a href="#">Contact</a></div>
-</section>
-
-<section class="overlay">
-    <div class="overlay-detail">
-        <p class="overlay-detail-title">Equipment Name</p>
-        <ul>
-            <li>Division:</li>
-            <li>Campus:</li>
-            <li>Material:</li>
-            <li>Characterization:</li>
-            <li>Overview:</li>
-            <li>Contact: </li>
-            <li>Link to Manual: </li>
-        </ul>
-    </div>
-</section>
-
-<section class="top">
-    <div class=" top top-logo__wrapper">
-        <p class="top-logo__placeholder">Logo</p>
-    </div>
-
-    <div class="top top-search__wrapper">
-        <label for="top-searchBox" hidden>Search Box</label>
-    <input type="text" class="reset top-searchBox" id="top-searchBox" />
-    <button class="reset top-searchButton">Search</button>
-    </div>
-
-    <div class="top top-login_wrapper">
-        <button class="top-loginButton">Login</button>
-    </div>
-
-</section>
-
-
-<!--------------- dropdown --------------->
-
-{% for cat in cats %}
-<section class="catDropdown" id="cat{{loop.index}}Dropdown">
-    {% if cat.description %}
-    <p class="catDropdown__description">{{ cat.description }}</p>
-    {% endif %}
-    <div class="catDropdown__bwrapper">
-        {% for k, bucket in cat.buckets.items() %}
-        <div class="catDropdown__bucket">
-            <p class="catDropdown__text">{{ bucket.name }}</p>
-            {% for choice in bucket.choices %}
-            <input type="radio" class="catDropdown__item js__catSelectionItem"
-                   id="{{ cat.name }}:{{ bucket.name }}-{{ choice }}">{{ choice }}<br>
-            {% endfor %}
-        </div>
-        {% endfor %}
-    </div>
-</section>
-{% endfor %}
-
-
-<section class="sidebar">
-
-    <div class="catNav sidebar-catNav__wrapper">
-        {% for cat in cats %}
-        <div class="catNav sidebar-catNav__cat" id="cat{{loop.index}}">
-            <p class="sidebar-catNav__text">{{ cat.name }}</p>
-        </div>
-        {% endfor %}
-    </div>
-
-    <div class="sidebar-refine__wrapper ">
-
-    </div>
-
-</section>
-
-<section class="result">
-    <div class="result-description__wrapper">
-        <p class="result-description__text">9 results found.</p>
-    </div>
-
-    {% for equipment in demos %}
-    <div class="result-block">
-        <div class="result-block-thumbnail__wrapper">
-            {% if 'link_image' in equipment %}
-            <img src="{{ equipment.link_image }}"
-                 alt="{{ equipment.name }}"
-                 class="result-block-thumbnail__img">
-            {% else %}
-            <img src="{{ url_for('static', filename='assets/image_placeholder.jpg') }}"
-                 alt="Placeholder Image"
-                 class="result-block-thumbnail__img">
-            {% endif %}
-        </div>
-        <div class="result-block-name__wrapper">
-            <a href="#" class="result-block-name__link">{{ equipment.name }}</a>
-        </div>
-    </div>
-    {% endfor %}
-
-
-
-<!--    <div class="result-pages__wrapper">-->
-<!--        <ul class="result-pages__list">-->
-<!--            <li class="result-pages__item">1</li>-->
-<!--            <li class="result-pages__item">2</li>-->
-<!--            <li class="result-pages__item">3</li>-->
-<!--            <li class="result-pages__item">4</li>-->
-<!--            <li class="result-pages__item">5</li>-->
-<!--            <li class="result-pages__item">...</li>-->
-<!--            <li class="result-pages__item">9</li>-->
-<!--        </ul>-->
-<!--    </div>-->
-</section>
-
-<section class="footer">
-    <div class="footer-placeholder">Footer</div>
-</section>
-<script src="{{ url_for('static', filename='demo.js') }}"></script>
-
-</body>
-</html>
