--- web/templates/edit.html
@@ -1,151 +1,0 @@
-{% extends 'base.html' %}
-
-{% block header %}
-    <h1 class="header-h1">{% block title %}Edit{% endblock %}</h1>
-{% endblock %}
-
-{% block script %}
-    <script src="{{ url_for('static', filename='edit.js') }}"></script>
-{% endblock %}
-
-
-
-
-{% block content %}
-    {% if not obj_id == 'new' %}
-        <p><a href="{{ url_for('edit.delete', obj_id=obj_id) }}">Delete</a></p>
-    {% endif %}
-
-    <form enctype="multipart/form-data" action="/edit/{{ obj_id }}" method="post">
-
-        <div id="editFields">
-            {% if obj_id == 'new' %}
-
-                Name: <input class="field required" type="text" name="name"><br>
-                Overview: <input class="field required" type="text" name="overview">
-
-                <div id="features">
-                    <p>Key Features: </p>
-                    <div id="feature1">
-                        <input class="field" type="text" name="feature1">
-                        <button class="delFieldButton" type="button" id="delFeature1">x</button>
-                    </div>
-                </div>
-                <button type="button" class="addField" id="addFeature">+</button>
-
-                <div id="applications">
-                    <p>Key Applications:</p>
-                    <div id="application1">
-                        <input class="field" type="text" name="application1">
-                        <button class="delFieldButton" type="button" id="delApplication1">x</button>
-                    </div>
-                </div>
-                <button type="button" class="addField" id="addApplication">+</button>
-
-                <div id="tags">
-                    <p>Tags:</p>
-                    <div id="tag1">
-                        <input class="field" type="text" name="tag1">
-                        <button class="delFieldButton" type="button" id="delTag1">x</button>
-                    </div>
-                </div>
-                <button type="button" class="addField" id="addTag">+</button>
-
-                <div id="documentation">
-                    <p>Link to Documentation: <input class="field" type="text" name="documentation"></p>
-                </div>
-
-                <div id="image">
-                    <p>Upload image (gif, jpg, png, jpeg)</p>
-                    <input type="file" id="imageField" name="image" accept="image/*">
-                    <button type="button" id="resetImage">Clear Upload</button>
-                </div>
-
-            {% else %}
-
-                <p>Name: <input class="field required" type="text" name="name" value="{{ result['name'] }}"></p>
-                <p>Overview: <input class="field required" type="text" name="overview" value="{{ result['overview'] }}"></p>
-
-                <div id="features">
-                    <p>Key Features: </p>
-                    {% if result['key_applications'] %}
-                        {% for feature in result['key_features'] %}
-                            <div id="feature{{ loop.index }}">
-                                <input class="field" type="text" name="feature{{ loop.index }}" value="{{ feature }}">
-                                <button class="delFieldButton" type="button" id="delFeature{{ loop.index }}">x</button>
-                                <br>
-                            </div>
-                        {% endfor %}
-                    {% else %}
-                        <div id="feature1">
-                            <input class="field" type="text" name="feature1">
-                            <button class="delFieldButton" type="button" id="delFeature1">x</button>
-                        </div>
-                    {% endif %}
-                </div>
-                <button type="button" class="addField" id="addFeature">+</button>
-
-                <div id="applications">
-                    <p>Key Applications:</p>
-                    {% if result['key_applications'] %}
-                        {% for application in result['key_applications'] %}
-                            <div id="application{{ loop.index }}">
-                                <input class="field" type="text" name="application{{ loop.index }}"
-                                       value="{{ application }}">
-                                <button class="delFieldButton" type="button" id="delApplication{{ loop.index }}">x
-                                </button>
-                            </div>
-                        {% endfor %}
-                    {% else %}
-                        <div id="application1">
-                            <input class="field" type="text" name="application1">
-                            <button class="delFieldButton" type="button" id="delApplication1">x</button>
-                        </div>
-                    {% endif %}
-                </div>
-                <button type="button" class="addField" id="addApplication">+</button>
-
-
-                <div id="tags">
-                    <p>Tags:</p>
-                    {% if result['tags'] %}
-                        {% for tag in result['tags'] %}
-                            <div id="tag{{ loop.index }}">
-                                <input class="field" type="text" name="tag{{ loop.index }}" value="{{ tag }}">
-                                <button class="delFieldButton" type="button" id="delTag{{ loop.index }}">x</button>
-                            </div>
-                        {% endfor %}
-                    {% else %}
-                        <div id="tag1">
-                            <input class="field" type="text" name="tag1">
-                            <button class="delFieldButton" type="button" id="delTag1">x</button>
-                        </div>
-                    {% endif %}
-                </div>
-                <button type="button" class="addField" id="addTag">+</button>
-
-                <div id="documentation">
-                    <p>Link to Documentation: <input class="field" type="text" name="documentation" value=""></p>
-                </div>
-
-                <div id="image">
-                    <p>Upload image (gif, jpg, png, jpeg)</p>
-                    {% if result['image'] %}
-                        <div id="imagePreview">
-                            <p>Uploaded: {{ result['image'] }}</p>
-                            <img src="{{ url_for('static', filename='user_uploads/thumbnails/'+result['image']) }}"
-                                 alt="{{ result['image'] }}">
-                        </div>
-                    {% endif %}
-                    <input type="file" id="imageField" name="image" accept="image/*"
-                           {% if result['image'] %} value="{{ result['image'] }}" {% endif %}>
-                    <button type="button" id="resetImage">Clear Upload</button>
-                </div>
-
-            {% endif %}
-        </div>
-        <br>
-
-        <input type="submit" id="submitField" value="Submit">
-    </form>
-{% endblock %}
